+++ src/hoiformat/gui.ts
@@ -1,8 +1,11 @@
+import { Token } from "./hoiparser";
 import { NumberLike, SchemaDef, positionSchema, Position, StringIgnoreCase } from "./schema";
 
 export interface Size {
     width: NumberLike;
     height: NumberLike;
+    x: NumberLike;
+    y: NumberLike;
 }
 
 export interface ComplexSize extends Size {
@@ -18,10 +21,8 @@ export interface Margin {
 
 export type Format = StringIgnoreCase<'left' | 'right' | 'up' | 'down' | 'center'>;
 export type Orientation = StringIgnoreCase<
-    'upper_left' | 'upper_right' | 'upper_center' |
-    'lower_left' | 'lower_right' | 'lower_center' |
-    'center_up' | 'center_down' | 'center_left' | 'center_right' | 'center_middle' | 'center' |
-    'left' | 'right' | 'up' | 'down' | 'top' | 'bottom'
+    'upper_left' | 'upper_right' | 'lower_left' | 'lower_right' |
+    'center_up' | 'center_down' | 'center_left' | 'center_right' | 'center'
 >;
 
 export interface Background {
@@ -33,6 +34,7 @@ export interface Background {
 
 export interface GuiTypes {
     containerwindowtype: ContainerWindowType[];
+    windowtype: ContainerWindowType[];
 }
 
 export interface ContainerWindowType {
@@ -44,10 +46,16 @@ export interface ContainerWindowType {
     margin: Margin;
     background: Background;
     containerwindowtype: ContainerWindowType[];
+    windowtype: ContainerWindowType[];
     gridboxtype: GridBoxType[];
     icontype: IconType[];
     instanttextboxtype: InstantTextBoxType[];
+    textboxtype: InstantTextBoxType[];
+    buttontype: ButtonType[];
+    checkboxtype: ButtonType[];
+    guibuttontype: ButtonType[];
     _index: number;
+    _token: Token;
 }
 
 export interface GridBoxType {
@@ -59,6 +67,7 @@ export interface GridBoxType {
     slotsize: Size;
     format: Format;
     _index: number;
+    _token: Token;
 }
 
 export interface IconType {
@@ -69,7 +78,9 @@ export interface IconType {
     spritetype: string;
     quadtexturesprite: string;
     frame: number;
+    scale: number;
     _index: number;
+    _token: Token;
 }
 
 export interface InstantTextBoxType {
@@ -82,7 +93,24 @@ export interface InstantTextBoxType {
     font: string;
     text: string;
     format: Format;
+    vertical_alignment: string;
     _index: number;
+    _token: Token;
+}
+
+export interface ButtonType {
+    name: string;
+    orientation: Orientation;
+    position: Position;
+    spritetype: string;
+    quadtexturesprite: string;
+    frame: number;
+    text: string;
+    buttontext: string;
+    buttonfont: string;
+    scale: number;
+    _index: number;
+    _token: Token;
 }
 
 export interface GuiFile {
@@ -92,6 +120,8 @@ export interface GuiFile {
 const sizeSchema: SchemaDef<Size> = {
     width: "numberlike",
     height: "numberlike",
+    x: "numberlike",
+    y: "numberlike",
 };
 
 const marginSchema: SchemaDef<Margin> = {
@@ -131,6 +161,7 @@ const iconTypeSchema: SchemaDef<IconType> = {
     spritetype: "string",
     quadtexturesprite: "string",
     frame: "number",
+    scale: "number",
 };
 
 const instantTextBoxTypeSchema: SchemaDef<InstantTextBoxType> = {
@@ -143,6 +174,20 @@ const instantTextBoxTypeSchema: SchemaDef<InstantTextBoxType> = {
     format: "stringignorecase",
     font: "string",
     text: "string",
+    vertical_alignment: "string",
+};
+
+const buttonTypeSchema: SchemaDef<ButtonType> = {
+    name: "string",
+    spritetype: "string",
+    quadtexturesprite: "string",
+    position: positionSchema,
+    orientation: "stringignorecase",
+    frame: "number",
+    text: "string",
+    buttontext: "string",
+    buttonfont: "string",
+    scale: "number",
 };
 
 const containerWindowTypeSchema: SchemaDef<ContainerWindowType> = {
@@ -157,6 +202,10 @@ const containerWindowTypeSchema: SchemaDef<ContainerWindowType> = {
         _innerType: undefined as any,
         _type: "array",
     },
+    windowtype: {
+        _innerType: undefined as any,
+        _type: "array",
+    },
     gridboxtype: {
         _innerType: gridBoxTypeSchema,
         _type: "array",
@@ -168,13 +217,34 @@ const containerWindowTypeSchema: SchemaDef<ContainerWindowType> = {
     instanttextboxtype: {
         _innerType: instantTextBoxTypeSchema,
         _type: "array",
-    }
+    },
+    textboxtype: {
+        _innerType: instantTextBoxTypeSchema,
+        _type: "array",
+    },
+    buttontype: {
+        _innerType: buttonTypeSchema,
+        _type: "array",
+    },
+    checkboxtype: {
+        _innerType: buttonTypeSchema,
+        _type: "array",
+    },
+    guibuttontype: {
+        _innerType: buttonTypeSchema,
+        _type: "array",
+    },
 };
 
 containerWindowTypeSchema.containerwindowtype._innerType = containerWindowTypeSchema;
+containerWindowTypeSchema.windowtype._innerType = containerWindowTypeSchema;
 
 const guiTypesSchema: SchemaDef<GuiTypes> = {
     containerwindowtype: {
+        _innerType: containerWindowTypeSchema,
+        _type: "array",
+    },
+    windowtype: {
         _innerType: containerWindowTypeSchema,
         _type: "array",
     },