+++ src/util/hoi4gui/common.ts
@@ -1,5 +1,5 @@
 import { NumberLike, Position, HOIPartial } from "../../hoiformat/schema";
-import { NumberSize, UserError } from "../common";
+import { NumberSize } from "../common";
 import { StyleTable } from '../styletable';
 import { Orientation, ComplexSize, Size, Margin } from "../../hoiformat/gui";
 
@@ -12,6 +12,7 @@ export interface RenderCommonOptions {
     id?: string;
     classNames?: string;
     styleTable: StyleTable;
+    enableNavigator?: boolean;
 }
 
 export function normalizeNumberLike(value: NumberLike, parentValue: number, subtractValue?: number): number;
@@ -24,84 +25,65 @@ export function normalizeNumberLike(value: NumberLike | undefined, parentValue: 
 
     switch (value._unit) {
         case '%': return value._value / 100.0 * parentValue;
-        case '%%': return Math.max(0, value._value / 100.0 * parentValue - subtractValue);
+        case '%%': return value._value / 100.0 * parentValue - subtractValue;
         default: return value._value;
     }
 }
 
 const offsetMap: Record<Orientation['_name'], { x: number, y: number }> = {
     'upper_left': { x: 0, y: 0 },
-    'upper_center': { x: 0.5, y: 0 },
     'upper_right': { x: 1, y: 0 },
     'lower_left': { x: 0, y: 1 },
-    'lower_center': { x: 0.5, y: 1 },
     'lower_right': { x: 1, y: 1 },
     'center_up': { x: 0.5, y: 0 },
     'center_down': { x: 0.5, y: 1 },
     'center_left': { x: 0, y: 0.5 },
     'center_right': { x: 1, y: 0.5 },
-    'center_middle': { x: 0.5, y: 0.5 },
     'center': { x: 0.5, y: 0.5 },
-    'left': { x: 0, y: 0.5 },
-    'right': { x: 1, y: 0.5 },
-    'up': { x: 0.5, y: 0 },
-    'down': { x: 0.5, y: 1 },
-    'top': { x: 0.5, y: 0 },
-    'bottom': { x: 0.5, y: 1 },
 };
 
-export function calculateStartLength(pos: NumberLike | undefined, size: NumberLike | undefined, parentSize: number, orientationFactor: number, origoFactor: number): [number, number] {
-    if (size?._unit !== '%%') {
-        // length mode
-        const baseStart = (normalizeNumberLike(pos, parentSize) ?? 0) + parentSize * orientationFactor;
-        const length = Math.max(0, normalizeNumberLike(size, parentSize) ?? 0);
-        const realStart = baseStart - length * origoFactor;
-        return [realStart, length];
-
-    } else {
-        // end mode
-        const baseStart = (normalizeNumberLike(pos, parentSize) ?? 0) + parentSize * orientationFactor;
-        const end = normalizeNumberLike(size, parentSize);
-
-        // not valid
-        if (origoFactor === 1) {
-            return [baseStart, 0];
-        }
-
-        // resolved from: realStart = baseStart - Math.max(0, end - realStart) * origoFactor
-        const realStart = (baseStart - end * origoFactor) / (1 - origoFactor);
-        const length = Math.max(0, end - realStart);
-        return [realStart, length];
+export function calculateStartLength(pos: NumberLike | undefined, size: NumberLike | undefined, parentSize: number, orientationFactor: number, origoFactor: number, scale: number): [number, number] {
+    let posValue = normalizeNumberLike(pos, parentSize) ?? 0;
+    let length = (normalizeNumberLike(size, parentSize) ?? 0) * scale;
+    if (size?._unit === '%%') {
+        length = length - posValue;
     }
+    if (length < 0) {
+        length = length + parentSize;
+    }
+
+    const start = posValue + parentSize * orientationFactor - length * origoFactor;
+    if (size?._unit === '%%' || (size?._value ?? 0) < 0) {
+        let end = normalizeNumberLike(size, parentSize) ?? 0;
+        if (end < 0) {
+            end = end + parentSize;
+        }
+        length = Math.max(0, end - start);
+    }
+
+    return [start, length];
 }
 
 export function calculateBBox(
-    {orientation, origo, position, size}: {
+    {orientation, origo, position, size, scale}: {
         orientation?: Orientation,
         origo?: Orientation,
         position?: Partial<Position>,
         size?: HOIPartial<ComplexSize> | Partial<Size & {min: undefined}>,
+        scale?: number,
     },
     parentInfo: ParentInfo
 ): [number, number, number, number, Orientation['_name']] {
-    const myOrientation = orientation?._name ?? parentInfo.orientation;
+    const myOrientation = orientation?._name ?? 'upper_left';
     const parentSize = parentInfo.size;
-    const orientationFactor = offsetMap[myOrientation];
-    const origoFactor = offsetMap[origo?._name ?? 'upper_left'];
+    const orientationFactor = offsetMap[myOrientation] ?? offsetMap['upper_left'];
+    const origoFactor = offsetMap[origo?._name ?? 'upper_left'] ?? offsetMap['upper_left'];
 
-    if (!orientationFactor) {
-        throw new UserError('Unknown orientation value: ' + myOrientation);
-    }
+    let [x, width] = calculateStartLength(position?.x, getWidth(size), parentSize.width, orientationFactor.x, origoFactor.x, scale ?? 1);
+    let [y, height] = calculateStartLength(position?.y, getHeight(size), parentSize.height, orientationFactor.y, origoFactor.y, scale ?? 1);
 
-    if (!origoFactor) {
-        throw new UserError('Unknown orientation value: ' + origo?._name);
-    }
-
-    let [x, width] = calculateStartLength(position?.x, size?.width, parentSize.width, orientationFactor.x, origoFactor.x);
-    let [y, height] = calculateStartLength(position?.y, size?.height, parentSize.height, orientationFactor.y, origoFactor.y);
-
-    const minWidth = normalizeNumberLike(size?.min?.width, parentSize.width, x) ?? width;
-    const minHeight = normalizeNumberLike(size?.min?.height, parentSize.height, y) ?? height;
+    const minWidth = normalizeNumberLike(getWidth(size?.min), parentSize.width, x) ?? width;
+    const minHeight = normalizeNumberLike(getHeight(size?.min), parentSize.height, y) ?? height;
     width = Math.max(minWidth, width);
     height = Math.max(minHeight, height);
 
@@ -122,4 +104,12 @@ export function removeHtmlOptions<T>(options: T): { [K in Exclude<keyof T, 'id' 
     delete result['id'];
     delete result['classNames'];
     return result;
+}
+
+export function getWidth(size?: Partial<Size>): NumberLike | undefined {
+    return size?.width ?? size?.x;
+}
+
+export function getHeight(size?: Partial<Size>): NumberLike | undefined {
+    return size?.height ?? size?.y;
 }