+++ src/util/hoi4gui/icon.ts
@@ -20,5 +20,21 @@ export async function renderIcon(icon: HOIPartial<IconType>, parentInfo: ParentI
         y -= image.height / 2;
     }
 
-    return renderSprite({x, y}, image, image, icon.frame ?? 0, options);
+    const scale = icon.scale ?? 1;
+
+    return `<div
+    start="${icon._token?.start}"
+    end="${icon._token?.end}"
+    class="
+        ${options.enableNavigator ? 'navigator navigator-highlight' : ''}
+        ${options.styleTable.style('positionAbsolute', () => `position: absolute;`)}
+        ${options.styleTable.oneTimeStyle('icon', () => `
+            left: ${x}px;
+            top: ${y}px;
+            width: ${image.width * scale}px;
+            height: ${image.height * scale}px;
+        `)}
+    ">
+        ${renderSprite({x: 0, y: 0}, image, image, icon.frame ?? 0, scale, options)}
+    </div>`;
 }